cleanup: move dbmanager from internal to pkg
