+++ pkg/dbmanager/types.go
